requirements.txt
PyJWT===2.1.0
asyncio
aiohttp
aiofiles
uvloop